# services/api-gateway/docs/swagger.yaml
definitions:
  errdetails.BadRequest:
    properties:
      field_violations:
        description: Describes all violations in a client request.
        items:
          $ref: '#/definitions/errdetails.BadRequest_FieldViolation'
        type: array
    type: object
  errdetails.BadRequest_FieldViolation:
    properties:
      description:
        description: A description of why the request element is bad.
        type: string
      field:
        description: "A path that leads to a field in the request body. The value
          will be a\nsequence of dot-separated identifiers that identify a protocol
          buffer\nfield.\n\nConsider the following:\n\n\tmessage CreateContactRequest
          {\n\t  message EmailAddress {\n\t    enum Type {\n\t      TYPE_UNSPECIFIED
          = 0;\n\t      HOME = 1;\n\t      WORK = 2;\n\t    }\n\n\t    optional string
          email = 1;\n\t    repeated EmailType type = 2;\n\t  }\n\n\t  string full_name
          = 1;\n\t  repeated EmailAddress email_addresses = 2;\n\t}\n\nIn this example,
          in proto `field` could take one of the following values:\n\n  - `full_name`
          for a violation in the `full_name` value\n  - `email_addresses[1].email`
          for a violation in the `email` field of the\n    first `email_addresses`
          message\n  - `email_addresses[3].type[2]` for a violation in the second
          `type`\n    value in the third `email_addresses` message.\n\nIn JSON, the
          same values are represented as:\n\n  - `fullName` for a violation in the
          `fullName` value\n  - `emailAddresses[1].email` for a violation in the `email`
          field of the\n    first `emailAddresses` message\n  - `emailAddresses[3].type[2]`
          for a violation in the second `type`\n    value in the third `emailAddresses`
          message."
        type: string
      localized_message:
        allOf:
        - $ref: '#/definitions/errdetails.LocalizedMessage'
        description: |-
          Provides a localized error message for field-level errors that is safe to
          return to the API consumer.
      reason:
        description: |-
          The reason of the field-level error. This is a constant value that
          identifies the proximate cause of the field-level error. It should
          uniquely identify the type of the FieldViolation within the scope of the
          google.rpc.ErrorInfo.domain. This should be at most 63
          characters and match a regular expression of `[A-Z][A-Z0-9_]+[A-Z0-9]`,
          which represents UPPER_SNAKE_CASE.
        type: string
    type: object
  errdetails.LocalizedMessage:
    properties:
      locale:
        description: |-
          The locale used following the specification defined at
          https://www.rfc-editor.org/rfc/bcp/bcp47.txt.
          Examples are: "en-US", "fr-CH", "es-MX"
        type: string
      message:
        description: The localized error message in the above locale.
        type: string
    type: object
  github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LocalSession:
    properties:
      sid:
        type: string
      uid:
        type: string
    type: object
  github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LoginErrorResponse:
    properties:
      message:
        type: string
      type:
        $ref: '#/definitions/github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LoginErrorResponseType'
      user_id:
        type: string
    required:
    - message
    - type
    type: object
  github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LoginErrorResponseType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    format: int32
    type: integer
    x-enum-varnames:
    - LoginErrorResponseTypeUserDeleted
    - LoginErrorResponseTypeBadCredentials
    - LoginErrorResponseTypeEmailNotVerified
    - LoginErrorResponseTypeBadBypassCode
    - LoginErrorResponseTypeLoginFromNewDevice
    - LoginErrorResponseTypeUserInBlacklist
  github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LoginInput:
    properties:
      bypass_code:
        example: 1A2B3C4D5E6F
        type: string
      device_trust_token:
        example: skdjfsdfsdifsdfsdf123
        type: string
      email:
        example: makc-dgek@mail.ru
        format: email
        type: string
      ip:
        example: 19.56.186.122
        format: ip4
        type: string
      password:
        type: string
      user_agent:
        example: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36
          (KHTML, like Gecko) Chrome/143.0.0.0 Safari/537.36
        type: string
    required:
    - email
    - ip
    - password
    - user_agent
    type: object
  httputils.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  sessionspb.CreatedTrustToken:
    properties:
      trust_token_exp_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      trust_token_value:
        type: string
    required:
    - trust_token_value
    type: object
  sessionspb.RefreshSessionRequest:
    properties:
      ip:
        example: 19.56.186.122
        format: ip4
        type: string
      refresh_token:
        type: string
      user_agent:
        example: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36
          (KHTML, like Gecko) Chrome/143.0.0.0 Safari/537.36
        type: string
    required:
    - ip
    - refresh_token
    - user_agent
    type: object
  sessionspb.SessionCreated:
    properties:
      access_exp_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      access_token:
        type: string
      refresh_exp_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      refresh_token:
        type: string
      trust_token:
        $ref: '#/definitions/sessionspb.CreatedTrustToken'
    required:
    - access_exp_time
    - access_token
    - refresh_exp_time
    - refresh_token
    type: object
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. This field is
          the nanosecond portion of the duration, not an alternative to seconds.
          Negative second values with fractions must still have non-negative nanos
          values that count forward in time. Must be between 0 and 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must
          be between -315576000000 and 315576000000 inclusive (which corresponds to
          0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z).
        type: integer
    type: object
  userspb.CreateUserRequest:
    properties:
      email:
        example: makc-dgek@mail.ru
        format: email
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  userspb.CreateUserResponse:
    properties:
      id:
        format: uuid
        type: string
    required:
    - id
    type: object
  userspb.EmailVerifiedSuccess:
    properties:
      message:
        type: string
    type: object
  userspb.GenderValue:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-varnames:
    - GenderValue_GENDER_UNSPECIFIED
    - GenderValue_GENDER_UNSELECTED
    - GenderValue_GENDER_MALE
    - GenderValue_GENDER_FEMALE
  userspb.Profile:
    properties:
      avatar_url:
        example: /avatars/photo.jpg
        type: string
      birth_date:
        example: "2001-11-23"
        type: string
      first_name:
        example: Ivan
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/userspb.GenderValue'
        description: |-
          GENDER_UNSPECIFIED = 0 Reserved for Proto, not a valid value
          GENDER_UNSELECTED = 1 User has not selected a gender
          GENDER_MALE = 2 Represents male gender
          GENDER_FEMALE = 3 Represents female gender
        example: 1
      last_name:
        example: Ivanov
        type: string
      middle_name:
        example: Olegovich
        type: string
      updated_at:
        $ref: '#/definitions/timestamppb.Timestamp'
    required:
    - gender
    - updated_at
    type: object
  userspb.SendNewEmailVerifyLinkRequest:
    properties:
      user_id:
        type: string
    type: object
  userspb.UpdateUserProfileData:
    properties:
      avatar_url:
        example: /avatars/photo.jpg
        type: string
      birth_date:
        example: "2001-11-23"
        type: string
      first_name:
        example: Ivan
        type: string
      gender_value:
        allOf:
        - $ref: '#/definitions/userspb.GenderValue'
        description: |-
          GENDER_UNSPECIFIED = 0 Reserved for Proto, not a valid value
          GENDER_UNSELECTED = 1 User has not selected a gender
          GENDER_MALE = 2 Represents male gender
          GENDER_FEMALE = 3 Represents female gender
        example: 1
      last_name:
        example: Ivanov
        type: string
      middle_name:
        example: Olegovich
        type: string
    type: object
  userspb.User:
    properties:
      created_at:
        $ref: '#/definitions/timestamppb.Timestamp'
      email:
        example: makc-dgek@mail.ru
        format: email
        type: string
      id:
        format: uuid
        type: string
      phone:
        example: "+79517735133"
        type: string
      profile:
        $ref: '#/definitions/userspb.Profile'
      updated_at:
        $ref: '#/definitions/timestamppb.Timestamp'
    required:
    - created_at
    - email
    - id
    - profile
    - updated_at
    type: object
  userspb.VerifyEmailLinkExpired:
    properties:
      message:
        type: string
      user_id:
        type: string
    type: object
  userspb.VerifyEmailLinkNotFound:
    properties:
      message:
        type: string
    type: object
info:
  contact:
    email: f3d0t0tvmax@yandex.ru
    name: Fedotv Maxim (developer)
  description: Swagger Documentation for API Gateway of microservices shop project
  title: Swagger Documentation for microservices shop API Gateway
  version: "1.0"
paths:
  /customers/session/check:
    get:
      consumes:
      - application/json
      description: Checking session auth for protected client routing
      parameters:
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LocalSession'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check session auth
      tags:
      - customers
  /customers/session/login:
    post:
      consumes:
      - application/json
      description: Login in account
      parameters:
      - description: Dto for login in account
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LoginInput'
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sessionspb.SessionCreated'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errdetails.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LoginErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Login in account
      tags:
      - customers
  /customers/session/refresh-session:
    post:
      consumes:
      - application/json
      description: Refresh session by refresh token from headers
      parameters:
      - description: Refresh session dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/sessionspb.RefreshSessionRequest'
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sessionspb.SessionCreated'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errdetails.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fedotovmax_microservices-shop_api-gateway_internal_domain.LoginErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Refresh session
      tags:
      - customers
  /customers/users:
    post:
      consumes:
      - application/json
      description: Create new user account
      parameters:
      - description: Create user account with body dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/userspb.CreateUserRequest'
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/userspb.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errdetails.BadRequest'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Create user account
      tags:
      - customers
  /customers/users/profile:
    get:
      consumes:
      - application/json
      description: Get user base profile
      parameters:
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userspb.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errdetails.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - customers
    patch:
      consumes:
      - application/json
      description: Update user profile
      parameters:
      - description: Update user profile with body dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/userspb.UpdateUserProfileData'
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errdetails.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - customers
  /customers/users/send-new-verify-email-link:
    patch:
      consumes:
      - application/json
      description: Send new email verification link on demand
      parameters:
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      - description: Send new email verification link with body dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/userspb.SendNewEmailVerifyLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errdetails.BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Send new email verify link
      tags:
      - customers
  /customers/users/verify-email/{link}:
    get:
      consumes:
      - application/json
      description: Verify email by verification link from user email address
      parameters:
      - description: Locale
        in: header
        name: X-Request-Locale
        type: string
      - description: Verification link
        in: path
        name: link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userspb.EmailVerifiedSuccess'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/userspb.VerifyEmailLinkExpired'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/userspb.VerifyEmailLinkNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Verify email
      tags:
      - customers
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer [ваш JWT токен]"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"